# openapi.yaml
openapi: 3.0.3
info:
  title: Last Price
  description: |
    Server for conducting A/B testing experiments on pricing strategies using Last Price integration.
    
    This server provides endpoints for:
    - Assigning users to experiment variants
    - Tracking pricing page views and conversions
    - Processing Paid.ai webhook events
    - Analyzing experiment results
    
    **Architecture**: The system uses deterministic SHA-256 hashing to ensure consistent user-to-variant mapping,
    emits signals to Paid.ai's Signals API for event tracking, and provides real-time analytics.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Retrieve an order
      description: Retrieves an order by ID from Last Price
      operationId: getLastPriceOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceOrder'

    delete:
      tags:
        - Orders
      summary: Delete an order
      description: Deletes an order from Last Price
      operationId: deleteLastPriceOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
  /orders/{orderId}/lines:
    put:
      tags:
        - Orders
      summary: Update order lines
      description: Updates order lines in Last Price
      operationId: updateLastPriceOrderLines
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lines:
                  type: array
                  items:
                    $ref: '#/components/schemas/LastPriceOrderLineInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LastPriceOrderLine'
  /orders/{orderId}/activate:
    post:
      tags:
        - Orders
      summary: Activate an order
      description: Activates an order in Last Price
      operationId: activateLastPriceOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceOrder'
  /usage/signals:
    post:
      tags:
        - Usage
      summary: Record agent signal
      description: Records a usage signal to Last Price
      operationId: recordLastPriceSignal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                signal:
                  $ref: '#/components/schemas/LastPriceAgentSignal'
      responses:
        '201':
          description: Signal recorded
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /invoices:
    get:
      tags:
        - Invoices
      summary: List all invoices
      description: Retrieves all invoices from Last Price
      operationId: listLastPriceInvoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, sent, paid, overdue, cancelled]
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LastPriceInvoice'

    post:
      tags:
        - Invoices
      summary: Create an invoice
      description: Creates a new invoice in Last Price
      operationId: createLastPriceInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LastPriceInvoiceInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceInvoice'
...
  /payments:
    get:
      tags:
        - Payments
      summary: List all payments
      description: Retrieves all payments from Last Price
      operationId: listLastPricePayments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, failed, refunded]
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LastPricePayment'
...
components:
  schemas:
...
    LastPriceOrderLine:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        agentId:
          type: string
          format: uuid
        agentExternalId:
          type: string
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        totalAmount:
          type: number
        billedAmountWithoutTax:
          type: number
        billedTax:
          type: number
        totalBilledAmount:
          type: number
    LastPriceOrder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        accountId:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        totalAmount:
          type: number
        estimatedTax:
          type: number
        billedAmountNoTax:
          type: number
        billedTax:
          type: number
        totalBilledAmount:
          type: number
        pendingBillingAmount:
          type: number
        creationState:
          type: string
        orderLines:
          type: array
          items:
            $ref: '#/components/schemas/LastPriceOrderLine'
...
    LastPriceOrderLineInput:
      type: object
      properties:
        agentExternalId:
          type: string
        agentId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    LastPriceAgentSignal:
      type: object
      required:
        - agentId
        - eventName
        - customerId
      properties:
        agentId:
          type: string
        eventName:
          type: string
        customerId:
          type: string
        data:
          type: object
          additionalProperties: true

    # ============ Invoice Schemas ============
    LastPriceInvoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        invoiceNumber:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        issueDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number
        currency:
          type: string
          example: USD
        lineItems:
          type: array
          items:
            $ref: '#/components/schemas/LastPriceInvoiceLineItem'
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LastPriceInvoiceInput:
      type: object
      required:
        - customerId
        - lineItems
      properties:
        customerId:
          type: string
          format: uuid
        issueDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        lineItems:
          type: array
          items:
            $ref: '#/components/schemas/LastPriceInvoiceLineItem'
        notes:
          type: string
        currency:
          type: string
          default: USD
    LastPriceInvoiceLineItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        amount:
          type: number

    # ============ Payment Schemas ============
    LastPricePayment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        invoiceId:
          type: string
          format: uuid
        amount:
          type: number
        currency:
          type: string
          example: USD
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        paymentMethod:
          type: string
        transactionId:
          type: string
        processedAt:
          type: string
          format: date-time
        refundedAmount:
          type: number
        refundedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    # ============ Dispute Schemas ============
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request parameters
        message:
          type: string
          example: userId must be a non-empty string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token for Last Price API authentication
    webhookSignature:
      type: apiKey
      in: header
      name: x-paid-signature
      description: HMAC signature for webhook verification